requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
alembic
psycopg2-binary

python-jose[cryptography]
pwdlib[argon2]

python-dotenv

razorpay

pydantic
python-multipart
requests

